+++ src/pretalx/submission/forms/submission.py
@@ -12,6 +12,11 @@ from pretalx.submission.models import Submission, SubmissionStates
 
 
 class InfoForm(RequestRequire, PublicContent, forms.ModelForm):
+    additional_speaker = forms.EmailField(
+        label=_('Additional Speaker'),
+        help_text=_('If you have a co-speaker, please add their email address here, and we will invite them to create an account. If you have more than one co-speaker, you can add more speakers after finishing the submission process.'),
+        required=False,
+    )
     def __init__(self, event, **kwargs):
         self.event = event
         self.readonly = kwargs.pop('readonly', False)
@@ -35,6 +40,8 @@ class InfoForm(RequestRequire, PublicContent, forms.ModelForm):
                 self.fields['track'].queryset = event.tracks.all()
             elif instance and instance.state != SubmissionStates.SUBMITTED:
                 self.fields.pop('track')
+        if instance and instance.pk:
+            self.fields.pop('additional_speaker')
 
         self._set_submission_types(instance=instance)